+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 Version 0.19
 ------------
-- curl_get() function now includes timing, fold retcode into rc
-- syntax error fix :-(
+- curl_get() function now includes timing, fold retcode return field into rc
+- pg_datalinker now uses SELECT FOR UPDATE, less verbose
+- stupid syntax error fix :-(
 
 Version 0.18
 ------------