+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+Version 0.20
+------------
+
 Version 0.19
 ------------
 - curl_get() function now includes timing, fold retcode return field into rc